Base the compilation on the result of the linker front-end.
Instead of manipulating raw `ClassDef`s, we now first invoke the
standard linker front-end to get a set of `LinkedClass`es. This
way, we benefit from the existing reachability analysis and method
synthesis (such as default method bridges).

One drawback is that abstract methods are removed from
`LinkedClass`es, including in interfaces. We recover the ones we
need with an additional pre-processing pass.