Elementary infrastructure for strings.
We introduce the global type `$string` as a typed struct containing
a reference to an array of `i16`. We add support for loading string
constants, and for converting primitive booleans to strings.

This is the minimum amount of support required to remove
`java.lang.Boolean$` from the excluded list. Its `toString(b: bool)`
method performs `"" String_+ b`.

This encoding of strings is hopefully temporary. Ideally we will
want to use a Wasm `stringref` instead.