Add support for infinite `While(true, ...)` loops.
In the IR, a `While(BooleanLiteral(true), ...)` has a special
typing rule that gives it the type `nothing` (instead of the
typical `void`). Therefore, it must be translated in a way that
ends with `unreachable` in the wasm code, so that it can be used
where any type is expected, such as `i32`.

In the process, we improve the translation of non-infinite loops
so that we emit the `cond` only once.