Pass the expected type down the codegen methods.
`genTree` now takes an `expectedType`. It computes what kind of
tree it actually generates. If the generated type is such that the
wasm type is not a subtype of the expected wasm type, `genAdapt`
makes the adaptation.

Currently, the only real adaptation that we need is from any type
to `NoType` (i.e., `void`). Indeed, in the IR type system, `void`
is a supertype of all types, but it is not the case in the wasm
type system.