Implement method dispatch for hijacked classes.
When the type of the receiver of an `Apply` is an ancestor of any
hijacked class, it might contain a JS value instead of one of our
own objects. In that case, we must perform a more elaborate
dispatch that tests whether we have a true object, or whether we
must dispatch to a hijacked class method.

This change also surfaces that the methods of `j.l.Object` can
receive an arbitrary non-null value, i.e., a `(ref any)`. This has
to be reflected in vtables and itables, unfortunately.