fix: Add tests for virtual dispatch + fix bug for abstract class
We had sevelral problems with the vtable implementations

- We'd been generating a global vtable even for an abstract class, but it's not needed because we can't instantiate abstract class
- For calculating the vtable (both it's type and global instance), we filtered out the abstract methods. Therefore, we couldn't resolve a abstract method call from vtable.

For example,

```scala
class A extends B:
  def a = 1

class B extends C:
  def b: Int = 1
  override def c: Int = 1

abstract class C:
  def c: Int
```

The vtable type for C will be an empty table, because there's no concrete methods. Therefore, when we have `x: C`, `x.c` wond't resolve the implementation of `C` because vtable type doesn't have a slot for `c`.

The root cause is that we generated both of the following from one (in-memory) vtable object that doesn't have abstract methods.

- vtable type (for declaring the vtable type, and resolve methods by name at compile time), and
- global vtable instance (for method lookup at runtime)

The former should include abstract methods like `C.c`, and the former should not.