Implement support for `classOf`.
Including elementary support of methods of `java.lang.Class`.

We implement this in a way that is similar to the JavaScript
back-end: we have an internal `typeData` structure, which contains
the raw metadata that will be needed by a `java.lang.Class`
instance. It also contains a lazily evaluated pointer to the
unique `java.lang.Class` and the `typeData` of an array of that
type. If we need the `typeData` of a multi-dimensional array, we
follow the chain of `arrayOf` pointers.

The name is initially stored as an `(array u16)` so that it can be
initialized as a constant expression in globals. The `string`
value is lazily initialized from that raw data the first time it
is requested.

This machinery requires quite a bit of run-time helper functions.
These are hard-coded in `HelperFunctions.scala`. Again, this is
similar to the JS back-end, which defines a bunch of helper JS
functions in its `CoreJSLib.scala`.

The `typeData` of classes that have a `vtable` are inserted into
the `vtable` itself, making the latter a subtype of `typeData`.
This is done in anticipation of the support for `GetClass`: it
will be able to fetch the `vtable` field of the object as the
`typeData` from which to extract the `java.lang.Class` instance.
The `typeData` of other classes are stored as independent
globals.