test: Remove `@JSExportTopLevel` from test-suites functions
Previously, the resulting Wasm module contained all the code for every test-suite package,
which made debugging test results harder because
(1) the resulting Wasm module contained a lot of irrelevant code, and
(2) if one of the test-suites was compiled to invalid Wasm code, all the tests would be affected, rendering all modules unable to instantiate.

This issue caused because each Scala.js package in the test-suites containing
`@JSExportTopLevel`, which served as a root node for reachability analysis.
Thus, all the code under test-suites remained alive after linking.

To address this, this commit removes all `@JSExportTopLevel` from the test-suites.
Instead, we added a function call to the `ModuleInitializer` methods into
the start function of Wasm
If a test fails, Wasm will execute unreachable, causing instantiation to fail and the test to fail accordingly.