Fix #27: Implement isInstanceOf[interface]
https://github.com/tanishiking/scala-wasm/issues/27

Current implementation is temporal and we should switch to the
efficient type inclusion test implementation in future.

This impl test if the given expression is instance of the given
interface by walking through the itables of the Object, which takes O(N)
(where N = number of interfaces the expr implements)

Also, this procedure can't be extracted as a helper function in Wasm because the
immediate argument of `ref.test` is specific to the `testClassName`,
which result in a number of instructions everywhere at `isInstanceOf[interface]`.

see: https://github.com/tanishiking/scala-wasm/issues/27#issuecomment-2008252049