What if buildAndAddToContext doesn't receive the FunctionTypeName?
Currently, buildAndAddToContext automatically generate the name for the
WasmFunctionTypeName such as `fun_type___0`. Therefore, `clone` function
will have a FunctionTypeName `fun_type___x`.

On the other hand, when we invoke or create a struct that contains
the function reference to `clone` funtion, we need to know the function
type name, `fun_type___x`.
To know that name, we need to store the generated name in the context
and pass it around the caller of the clone function.