Make array values actual subclasses of java.lang.Object.
Previously, arrays such as `Array[Int]` were directly represented
as a `(ref null (array i32))`. This was of course efficient, but
not compatible with the universal representation of objects when
upcast to `anyref`.

We now represent them as instances of "array classes", which truly
extend `j.l.Object`. The array classes have a single field, which
is the underlying `(ref (array i32))` (or other type of array).

Moreover, all reference array types are reprensented by a single
struct type whose underlying is `(ref (array anyref))`. This is
necessary because reference array types are covariant in the IR,
like in JVM bytecode. Instances however have different vtables,
which still allows to distinguish them at run-time.