Fix #40: Implement Is/AsInstanceOf for array types.
`IsInstanceOf` is not yet correct for reference array types, as it
treats all of the same as if they were `Array[jl.Object]`. In order
to do the right thing here, we need more support of the subtyping
relationship in `typeData`s. However, what we have now should be
enough to support more real-world code, as instance tests to
reference types other than `Array[jl.Object]` are rare.