Encapsulate the scope management of `LocalName`s in WasmFunctionContext.
In the process, we change the strategy to unpack capture data structs
in closure functions. Previously, we unpacked all the captures at the
start of the function. Now, we unpack them only when needed.

A function context now remembers its current environment, which is a
map from `LocalName` to `VarStorage`. A `VarStorage` can either be
an actual `Local` variable, or a `StructField`. The latter is used
for capture data structs.

This protects the choice of `WasmLocalName`s entirely inside
`WasmFunctionContext`.