Fix #42: Implement consistent interface table indexing
This change ensures that the interface tables (itables) array for each class are consistently ordered:
Each interface is assigned a designated index, and its corresponding itable will always be placed at that index in the global itables array.

For example, consider the following Scala code:

```scala
trait A // index = 0
trait B // index = 1

class Foo extends A with B
class Bar extends B with A
```

Both `Foo` and `Bar` will have itables arrays with a length equal to the number of interfaces (A and B in this case).
The itable for `A` will be at index 0, and the itable for `B` will be at index 1, regardless of the order in which the traits are mixed into the class.