Optimize type inclusion test for interfaces to O(1)
This commit optimizes the type inclusion test for interfaces to  O(1).

Previously, the type inclusion test walking through the interface tables (itables) and returning true if an itable instance was found, resulting in O(N) time complexity.

The optimization leverages the consistent interface table indexing introduced in the previous commit.
Each interface now has a designated index, and its corresponding itable is always placed at that index in the global itables array for all classes that implement that interface.

With this change, the type inclusion test for an interface can be performed by looking up the designated index corresponding to the interface from the class's itables array, and checking if the entry at that index is non-null.
This operation is constant-time, O(1), as it involves a single array lookup and a null check.

For example, the instance test for the `Base1` interface is generated as follows:

```wasm
(func $instanceTest#testsuite.core.Base1__fun (type $fun_type___4___ty)
  (param $expr___local anyref) (result i32)
  (local $local___0___local (ref null $itable___arr))
  block $0 (result (ref null any))
    local.get $expr___local
    br_on_cast_fail $0 (ref null any) (ref $java.lang.Object___struct)
    struct.get $java.lang.Object___struct 1
    local.set $local___0___local
    ref.null any
    local.get $local___0___local
    br_on_null $0
    local.get $local___0___local
    i32.const 5 ;; index of Base1 interface
    array.get $itable___arr
    br_on_null $0
    i32.const 1
    return
  end
  drop
  i32.const 0
)

Although this optimization addresses the time complexity concern, the generated itables array is still sparse.
There's a room for further optimizations to compress the itables and reduce memory overhead.

Related https://github.com/tanishiking/scala-wasm/issues/49