Fix #44: Support non-native JS classes.
The handling of the IR Trees themselves are straightforward. The
real work happens in `WasmBuilder`, and in particular in
`genCreateJSClassFunction`. The main difficulty is that we need to
decompose the JS constructor into three pieces, because the call to
the super constructor must syntactically appear as such in the
JavaScript source of the constructor. There is a big comment in the
code that explains the decomposition.