Merge the `cli` project into the `tests` project.
The `cli` was only used to link the test suites to be run by the
`tests` project. We now directly integrate its logic inside the
test executor of `tests`.

In addition to cleaning things up, this improves error reporting
in case a suite fails to link: the error will be reported as part
of the test for that suite, rather than as a global failure ahead
of time. Failing to link one suite does not prevent the other ones
from succeeding anymore.

Also update the readme about how to run tests.