Clean up handling of `NoType` and `NothingType`.
These can only be used as "result types", instead of general
"value types". Moreover, when `NothingType` is used as the result
type of a Wasm instruction, it must be followed by an additional
`unreachable` statement to recover stack polymorphism.

We also remove `WasmNoType`, as it does not actually in Wasm.