Simplify our model of Wasm types.
The main change is that all reference types are grouped under a
single `case class WasmRefType(nullable, heapType)`. This includes
reference types that have shorthands.

Shorthands are still used when writing the text and binary formats.

We introduce new constructor functions for `WasmRefType` to make
it convenient to use.

We also simplify `WasmHeapType`. The `Func` case is removed in
favor of always using `Type`. The "abstract heap types", as they
are called in the spec, are moved one level up to be more
convenient to use.