Refactor the logic for writing immediates.
Previously, we had a generic way of write `WasmImmediate`s. Each
instruction registered a list of `WasmImmediate`s to be written
in order.

This was nice in theory, but falls short for some instructions
whose binary and/or text encodings vary depending on their
arguments. This is the case of `ref.cast`, `ref.test`, `br_on_cast`
and `br_on_cast_fail`.

We now use a different strategy. We do not generically handle
immediates per se. Instead, we handle categories of instructions
based on their "shape": the specific immediates they take. Most
instructions fall into a shape with a single immediate or none at
all, so we end up with only a few more cases to handle.

One significant benefit is that it is easier to manipulate the
cast/test instructions mentioned above. Another benefit will be
to remove the `WasmImmediate` wrappers alltogether, but we will
split that into a separate commit for easier reasoning.