Remove `WasmImmediate` wrappers on all the immediates.
Now that we dispatch on instruction shape rather than on immediate
types, we do not need those wrappers anymore.

There are four types left in the `WasmImmediate` object, namely
`LabelIdx`, `StructFieldIdx`, `BlockType` and `CatchClause`.
They are left there in this commit to keep the latter
straightforward. The will be refactored further in the following
commits.