Handle `asInstanceOf[jl.Number]`, which must not insert a cast.
`jl.Number` is the only non-`Object` *class* that is an ancestor of
hijacked classes. Previously the code special-cased `Object`, but
then did not work for `jl.Number`.

Now we handle all ancestors of hijacked classes first, whether they
are classes or interfaces. We remove the special-case for `Object`,
as it is not necessary anymore.