Rudimentary support for reflective calls
Added a `reflectiveProxies` field to `typeData`, this field contains an
array of struct with `methodId` of `i32`, and a reference to a
reflective proxy function.
When the compiler found an Apply to the reflective proxy, we lookup the
method to call from the `reflectiveProxies` field in the receiver's
`typeData` field, instead of normal table dispatching.