Remove the subtypes of `WasmTypeName`.
Instead, make `WasmTypeName` a final case class. Since they live in
the same namespace at the Wasm level, it does not really make sense
to separate them in different types of case classes, which would
compare as being in different namespaces at the Scala level.

Moreover, we were barely using the better Scala types to constrain
types in practice, so that did not seem to be useful either.

The `object`s for `WasmFunctionTypeName`, `WasmArrayTypeName` and
`WasmStructTypeName` remain as builders for well-known type names.