Always use 4 bytes to store byte lengths of subsections.
This allows to directly write to a unique target buffer. It will
also provide known global offsets to generate source maps without
major recomputation hoops.

The amount of bytes "lost" should be about 2xN where N is the
number of declared functions. That is because we can expect most
functions to have a size that should require 2 bytes to store. The
lengths of actual full *sections* won't matter as there are only a
dozen of them.