Build vtables and itables separately from resolving concrete implementations.
Previously, we put in `classInfo.methods` two different things:

* concrete implementations of methods, and
* abstract definitions required for dynamic calls.

Now, we separate these two categories, and store them respectively
in `resolvedMethodInfos` and `tableEntries`/`tableMethodInfos`.

We also fill in everything in a set order from the preprocessor,
instead of computing `VTable`s lazily.

The separation can lead to creating smaller vtables, as not all
public methods are called dynamically. It is possible for public
methods to only be called through `ApplyStatically` instead.
This effect will be hightened in the following commit, in which we
will generate `Apply`s as `ApplyStatically` when possible (when the
target method is never overridden).