Generate simpler code for Int and Long division and remainder.
Notably, when the rhs is a constant, we can emit straightforward
code without any branches. This is worth it because the divisor of
integer divisions and remainders is often constant in practice.

To handle the overflow case of `MinValue / -1`, we also remove one
branch by turning `lhs / -1` into `0 - lhs`. The previous codegen
used `if (lhs == MinValue) MinValue else lhs / -1` in the same
situation.