Compress interface tables with packed encoding
This commit implements the "packed encoding" algorithm for compressing
interface tables (itables). The algorithm is based on the paper "Efficient Type Inclusion Tests".
https://www.researchgate.net/publication/2438441_Efficient_Type_Inclusion_Tests
It compresses the itable by reusing itable indices for unrelated types, reducing the size of the itable array.
For scalajs-test-suite, the itable size is reduced from 413 to 47.

This change didn't reduce the size of wasm binary that much (14957kb to 14943kb), but it should be space efficient.

We chose packed encoding because the compression rate is good enough among the presented methods in the paper. While hierarchical encoding is slightly better compression rate, packed encoding is better in both compile-time and runtime performance in CPU and space.