Refactoring: Move `*Context` to `ir2wasm`.
We extract the Wasm-only parts into new classes `ModuleBuilder`
and `FunctionBuilder`. The parts that interact with the IR and/or
contain "decision-making" related to our back-end stay in
`WasmContext` and `WasmFunctionContext` and are moved to `ir2wasm`.

We make `WasmModule` fully immutable in the process, since the
mutable aspects are moved to `ModuleBuilder`.