Directly use FunctionBuilder when we do not need WasmExpressionBuilder.
This way, we use `WasmFunctionContext`+`WasmExpressionContext` only
when we need to translate user-defined IR trees. In that case, the
`WasmFunctionContext` is responsible for local name management:
encoding from IR local names and generation of synthetic local names.

Otherwise, when we generate Wasm code independent of IR trees, we
directly use `FunctionBuilder`. Local name management is left to
the user code.