Isolate the evaluation of jsSuperClass and nameTree's.
In general, they can contain local variable definitions. These must
not be defined in the same scope as the `createJSClass` function
body, since they could clash with each other and/or with the class
captures.

Instead, in the general case, we extract them in separate functions,
which we immediately call. For the common cases of string literals
and direct reference to a class capture, we avoid the extra function.

This treatment is similar to what the JS backend does, except the
latter defines the "separate functions" inline, as immediatly
invoked function expressions (IIFEs).