Use namespacing conventions in CoreWasmLib.
In the CoreWasmLib, we manipulate IR types and names, but not trees,
so we import the former two without namespacing.

We produce a lot of Wasm types and instructions, so those are
imported without namespacing either.

This is similar to the import strategy of the `CoreJSLib` in the
JS backend.