Simplify the codegen for `AsInstanceOf`.
Previously, we tried to avoid useless casts based on IR subtyping
rules. Now, we do it based on Wasm type equality. Moreover, if we
do need to cast, we only use a direct RefCast, rather than
`unboxOrNull` helpers for hijacked classes.

The new scheme has several advantages:

* It avoids more useless casts. For example, it avoids downcasts to
  interface types when the source type is not an ancestor of a
  hijacked class, since they all translate to `(ref null jl.Object)`.
* The code is much simpler.
* It avoids an expensive IR subtyping test, which would also not
  survive well if we get to make the backend incremental.

As is, some casts can be introduced when they were not necessary
before. However, such cases are rare, and will be truly gone when
we can enable the optimizer.