Better codegen for the clone functions.
Previously, we first created the struct with zero's, then assigned
fields from the source one by one.

We now load all the fields from the source onto the stack, then
directly use `struct.new` to allocate the target.

It might not necessarily be more efficient, but it definitely
results is smaller code.